some code tidy up
